Added rules to toolview.css to try and fix display of list items.  They are now displayed in block mode but the bullet is still missing.
Fixed bug in tool.py where the "all" link at the top of the tool page was
pointing to a drawer rather than the toolbox.

Fixed bug in the toolview page template where variables where misnamed.
This was causing errors to be displayed when relavance had content but
resources did not.